SMAX: Reorganized Designer DTOs
