+++ Modules/Devon4Net.Infrastructure.SmaxHcm/Dto/Designer/CreateDesignContainer/CreateDesignContainerResponseDto.cs
@@ -1,4 +1,4 @@
-﻿namespace Devon4Net.Infrastructure.SmaxHcm.Dto.Designer
+﻿namespace Devon4Net.Infrastructure.SmaxHcm.Dto.Designer.CreateDesignContainer
 {
     public class CreateDesignContainerResponseDto
     {
