+++ Modules/Devon4Net.Infrastructure.SmaxHcm/Dto/Designer/GetDesignTags/GetDesignTagsResponseDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Devon4Net.Infrastructure.SmaxHcm.Dto.Designer
+namespace Devon4Net.Infrastructure.SmaxHcm.Dto.Designer.GetDesignTags
 {
     public class GetDesignTagsResponseDto
     {
