+++ Modules/Devon4Net.Infrastructure.SmaxHcm/Handler/SmaxHcmHandler.cs
@@ -8,6 +8,9 @@ using Devon4Net.Infrastructure.Common.Options.SmaxHcm;
 using Devon4Net.Infrastructure.SmaxHcm.Common;
 using Devon4Net.Infrastructure.SmaxHcm.Dto.Catalog;
 using Devon4Net.Infrastructure.SmaxHcm.Dto.Designer;
+using Devon4Net.Infrastructure.SmaxHcm.Dto.Designer.CreateDesignContainer;
+using Devon4Net.Infrastructure.SmaxHcm.Dto.Designer.CreateDesignVersion;
+using Devon4Net.Infrastructure.SmaxHcm.Dto.Designer.GetDesignTags;
 using Devon4Net.Infrastructure.SmaxHcm.Dto.Login;
 using Devon4Net.Infrastructure.SmaxHcm.Dto.Offering;
 using Devon4Net.Infrastructure.SmaxHcm.Dto.Providers;
